Add audio_common for microphone node and sp in uint8

+++ catkin_ws/src/voice/src/voice_scripts/speech_process.py
@@ -52,7 +52,7 @@ def process_file(filename):
     nperseg = math.floor(fs*2/100)
     overlap = math.floor(nperseg/2)
     NFFT = 2*nperseg
-    data_signal = np.array(a,dtype=float)
+    data_signal = np.array(a,dtype=uint8)
     data_signal = apply_preemphasis_filter(data_signal)
     f, t, Zxx = apply_stft(data_signal, nperseg, overlap, fs, NFFT)
     filter_banks = apply_filter_banks(Zxx, fs, NFFT)
